feat: add stylesheet to layout and create .eleventy.js and styles.css files
This commit includes the addition of a link to a new stylesheet in mylayout.njk.
A new .eleventy.js config file has also been created to add a passthrough copy for the new styles.css file.
Furthermore, a new styles.css file has been added with a basic style rule for h1 tags.